Backend/requirements.txt
urllib3==1.25.9
pandas==1.4.3
requests==2.23.0
openpyxl==3.0.10
Flask==2.2.2
python-dotenv==0.21.0
